libnm: don't validate "connection.interface-name" from "nm-setting-infiniband.c"'s verify()
There should not be multiple places to validate the interface-name.

The check in "nm-setting-infiniband.c" is unnecessary and wrong.

It's unnecessary, because _nm_connection_verify() takes care to
first verify the NMSettingConnection instance.

It's wrong, because it does not check the property the same way as
NMSettingConnection does (e.g. it does not check for valid UTF-8).