dhcp/nettools: first clear source before emitting event in dhcp4_event_cb()
A similar patch was done on master, but here the situation is different.

I feel we should not allow for the possibility where we invoke an event
that might mess with the source id. In practice there was no problem.
But it feels cleaner to clear it first.

Fixes: 843d696e46d9 ('dhcp: clean source on dispatch failure')
(cherry picked from commit 0549351111ccfde9d6a404952c43234612028d13)