libnm: hide NMVpnPluginInfo structs from public API
The NMVpnPluginInfo class is not intended to be subclassed. An API that
allows to be subclassed needs to be designed in a certain manner for
that to be useful. NMVpnPluginInfo does not want to support that.

Only because a user technically could do that (as the structs were
in the public headers), it does not make it supported. Not everything
that is possible in C is guaranteed to work.

Also, of course there exist no users in practice that would rely on this.
So, hide the structs.

Also, this allows to embed the private data in the GObject struct
itself, which is useful for debugging and for performance.