libnm: make NMSettingVpn's foreach_item_helper() robust against changes
When we invoke the user's callback, be more robust about the user
modifying or unreferencing the NMSettingVpn, while iterating.

While it would be odd to modify the NMSettingVpn from inside the foreach
callback, we should behave consistently and sensibly.

That means, to ensure that the NMSettingVpn instance stays alive all
the time, that we don't crash, and that we always iterate over the
previously determined, planned list of keys.

Also, avoid some unnecessary string copies, like the clone of the first key.