auth-manager: watch PolicyKit name owner to emit changed signal when name owner appears
We shall emit a changed signal when PolicyKit changes. That implies also
emitting the signal when PolicyKit starts. Watch the name owner for
that.

Note that I don't manage to actually receive the "Changed" signal from
PolicyKit. That seems to be due to wrong D-Bus configuration which
prevents the "Changed" signal from PolicyKit to reach NetworkManager.
That is independent from this patch.

https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/merge_requests/469