tests: ignore valrind false positives when running meson tests
On newer versions, the test binaries where renamed so that meson and
autotools build the same executables. Hence, the workaround for
valgrind warnings works on newer versions, but not on this branch.

Instead of backporting the renaming of the build artefacts, adjust
the test runner to work with meson builds.